merged program for extracting frames and cell shapes
with the merged program (extract_frames_shapes), it is no longer necessary to extract images from the video files. Images are taken directly out of the movie stream and are analyzed. The analyzed files (i.e. the flatfield image and the result file) are stored in the same directory as the video file.