added new program (strain vs stress)
in the extract farmes sheapoes program, the minor and major axis is now stored in units of µm (to make it consistent with the radial position in the channel).

A new program (strain_vs_stress.py) for plotting strain versus stress and for analyzing the stress stiffening equation (for computing cell mechanical properties has been added to the repository.