added a nearest neighbour deconvolution
perform a nearest neighbour deconvolution on each image prior to performing the Canny edge detection. Nearest neighbour deconvolution essentially is a band-pass filter that supresses out-of-focus features.

In addition, the program only selects cells where the binary shape perimeter is not more than 5% (or 6%) longer than the perimeter of the ellipse, and the solidity of the object (percentage of binary pixels within the convex hull) is >98% (or 95%). Number is backets are increasing the number of analyzed cells by 2-fold without selecting too many unsuitable cells that may distort the results.